rsb/requirements.txt
Django==3.2.9
django-admin-material==0.1.0
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.6